+++ CHANGELOG.md
@@ -4,7 +4,11 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
 
+Features:
+- Add `isExternal` option to `LinkBox`.
+
 Fixes:
+- Change some styles on `k-LinkBox`.
 - Fix text-decoration in `k-Paragraph__link` and `k-BulletList__link`.
 
 ## [v4.2.0] - 2016-12-23